break up long towerse file into small files, combine with vertical cylinder
